+++ jest.config.ts
@@ -1,7 +1,7 @@
 import type { Config } from "@jest/types";
 
 const config: Config.InitialOptions = {
-  setupFilesAfterEnv: ["./jest.setup.ts"],
+  setupFilesAfterEnv: ["./jest.setup.tsx"],
   testEnvironment: "jsdom",
   verbose: true,
   moduleNameMapper: {