+++ packages/react/src/atomic/Chip/src/chip.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { withA11y } from "@storybook/addon-a11y";
-import { BoxPackedIcon } from "@tiendanube/icons";
+import { BoxPackedIcon } from "@nimbus-ds/icons";
 
 import { Chip } from "./Chip";