+++ packages/react/src/atomic/Button/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The Button component allows us to initiate actions, make state or page changes.
 
+## 2023-02-14 `2.3.0`
+
+### 🎉 New features
+
+- Removed external dependency from `@tiendanube/icons` package to now use internal `@nimbus-ds/icons` package. ([#94](https://github.com/TiendaNube/nimbus-design-system/pull/#94) by [@juniorconquista](https://github.com/juniorconquista))
+
+### 📚 3rd party library updates
+
+- Removed `@tiendanube/icons@0.3.1`. ([#94](https://github.com/TiendaNube/nimbus-design-system/pull/#94) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-02-13 `2.2.0`
 
 ### 🎉 New features